requirements.txt
aiogram>=3.4.0
openai==1.6.1
PyYAML==6.0.1
httpx>=0.27
PyPDF2>=3.0.0
schedule>=1.2.0
pytz>=2023.3
